# docs/swagger.yaml
basePath: /
...
host: localhost:8081
info:
  contact: {}
  description: |-
    RESTful API for a Hotel Booking System built with Go, Gin, GORM, PostgreSQL, and JWT authentication.

    Features include authentication, hotel management, room management, bookings, availability, reviews, favorites, and payments.
  title: Hotel Booking System API
  version: "1.0"
paths:
...
  /bookings/{id}/payment:
    post:
      consumes:
      - application/json
      description: Creates a payment for the authenticated user's booking.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payment data
        in: body
        name: request
        required: true
        schema:
          properties:
            payment_method:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create payment
      tags:
      - Payments
...
  /payments:
    get:
      description: Retrieves all payments. Admin and manager only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all payments
      tags:
      - Payments
  /payments/{id}:
    get:
      description: Retrieves a payment by ID. Customers can only access their own
        payments. Admins and managers can access any payment.
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get payment by ID
      tags:
      - Payments
  /payments/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates a payment status. Admin and manager only. Supported statuses
        are paid, failed, and refunded.
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payment status
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update payment status
      tags:
      - Payments
  /payments/my:
    get:
      description: Retrieves all payments belonging to the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my payments
      tags:
      - Payments
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"